Updated CSV parser to now parse and print all information within the CSV
This parser makes use of the apache.commons.csv library to parse the CSVfile. This way each line from the CSV is it's own record allowing us to in the future to filter and display only information related to say a specific time or other value.

Note: the apache.commons.csv needs to be project dependencies, The Zip file will be placed in GIT.

author: Ethan King